+++ scripts/lib/utilities/tokenUtils.js
@@ -411,6 +411,52 @@ function getLinearDistanceMoved(token) {
     }
     return token.measureMovementPath(waypoints).distance;
 }
+async function mountToken(rider, target, {vae, unhideActivities} = {}) {
+    if (!game.modules.get('Rideable')?.active) return false;
+    let effect = effectUtils.getEffectByIdentifier(rider.actor, 'mountedRider');
+    if (effect) await genericUtils.remove(effect);
+    let effectData = {
+        name: genericUtils.translate('CHRISPREMADES.Macros.Mount.MountedRider'),
+        img: 'icons/environment/creatures/horse-brown.webp',
+        origin: rider.actor.uuid,
+        flags: {
+            'chris-premades': {
+                mount: {
+                    target: target.document.uuid
+                }
+            },
+            dae: {
+                showIcon: true
+            }
+        }
+    };
+    effectUtils.addMacro(effectData, 'effect', ['mountedRider']);
+    effect = await effectUtils.createEffect(rider.actor, effectData, {identifier: 'mountedRider', rules: 'modern', vae, unhideActivities});
+    let targetEffectData = {
+        name: genericUtils.translate('CHRISPREMADES.Macros.Mount.MountedTarget'),
+        img: 'icons/environment/creatures/horse-brown.webp',
+        origin: target.actor.uuid,
+        flags: {
+            'chris-premades': {
+                mount: {
+                    rider: rider.document.uuid
+                }
+            },
+            dae: {
+                showIcon: true
+            }
+        }
+    };
+    effectUtils.addMacro(targetEffectData, 'effect', ['mountedTarget']);
+    await effectUtils.createEffect(target.actor, targetEffectData, {
+        identifier: 'mountedTarget',
+        interdependent: true,
+        parentEntity: effect,
+        rules: 'modern'
+    });
+    await game.Rideable.Mount([rider.document], target.document);
+    return true;
+}
 export let tokenUtils = {
     getDistance,
     checkCover,
@@ -428,5 +474,6 @@ export let tokenUtils = {
     grappleHelper,
     getMovementHitTokens,
     getLinearDistanceMoved,
-    isGrappledBy
+    isGrappledBy,
+    mountToken
 };